requirements.txt
numpy~=1.23.5
tensorflow~=2.12.0
pandas~=1.5.3
Pillow~=9.4.0
keras~=2.12.0
torch~=2.0.0
transformers~=4.28.0.dev0
datasets~=2.11.0
requests~=2.28.2
matplotlib~=3.7.1
python-decouple~=3.8